+++ CHANGELOG.rst
@@ -11,9 +11,10 @@ Change Log
 
 .. There should always be an "Unreleased" section for changes pending release.
 
-Unreleased
-**********
+2.0.0 - 2024-01-03
+******************
 * Add content cache
+* Integrate system prompt setting
 
 1.5.0 - 2023-10-18
 ******************